requirements.txt
# ============================================================================
# ST-CDGM Project - Complete Requirements
# Spatio-Temporal Causal Diffusion Generative Model for Climate Downscaling
# ============================================================================

# --------------------------------------------------------------------------
# Core Scientific Libraries
# --------------------------------------------------------------------------
numpy>=1.21.0,<2.0.0
pandas>=1.3.0
xarray>=2023.1.0
scipy>=1.7.0

# --------------------------------------------------------------------------
# PyTorch Ecosystem
# --------------------------------------------------------------------------
# Note: Pour CUDA, installez avec: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0

# --------------------------------------------------------------------------
# PyTorch Geometric (Graph Neural Networks)
# --------------------------------------------------------------------------
# Installation: pip install torch-geometric
# Ou avec wheels: pip install torch-geometric torch-scatter torch-sparse -f https://data.pyg.org/whl/torch-2.0.0+cu118.html
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17

# --------------------------------------------------------------------------
# Diffusion Models
# --------------------------------------------------------------------------
diffusers>=0.21.0
accelerate>=0.20.0
transformers>=4.30.0  # Requis par certains modèles diffusers

# --------------------------------------------------------------------------
# NetCDF and Climate Data Processing
# --------------------------------------------------------------------------
netcdf4>=1.6.0
h5netcdf>=1.1.0
xbatcher>=0.3.0  # Pour les batches temporels spatio-temporels
dask[complete]>=2023.1.0  # Pour le traitement distribué
zarr>=2.14.0  # Format de données optimisé pour ML (alternative à NetCDF)

# --------------------------------------------------------------------------
# Visualization
# --------------------------------------------------------------------------
matplotlib>=3.5.0
seaborn>=0.12.0

# --------------------------------------------------------------------------
# Configuration Management
# --------------------------------------------------------------------------
hydra-core>=1.3.0
omegaconf>=2.3.0

# --------------------------------------------------------------------------
# Testing
# --------------------------------------------------------------------------
pytest>=7.0.0
pytest-cov>=4.0.0

# --------------------------------------------------------------------------
# Utilities
# --------------------------------------------------------------------------
tqdm>=4.64.0  # Barres de progression
python-dateutil>=2.8.0
pytz>=2022.1

# --------------------------------------------------------------------------
# Optional: Jupyter and Interactive Analysis
# --------------------------------------------------------------------------
jupyter>=1.0.0
jupyterlab>=3.5.0
ipykernel>=6.15.0
ipywidgets>=8.0.0
notebook>=6.5.0

# --------------------------------------------------------------------------
# Optional: Additional XAI and Visualization
# --------------------------------------------------------------------------
plotly>=5.10.0  # Visualisations interactives
networkx>=2.8.0  # Visualisation de graphes causaux

# --------------------------------------------------------------------------
# Optional: Additional Data Formats and Optimizations
# --------------------------------------------------------------------------
webdataset>=0.2.0  # Alternative WebDataset format (optional, Phase 1.1)
pyg-lib>=0.1.0  # Optimisations PyTorch Geometric (optional, Phase 2.4)
k-diffusion>=1.0.0  # Pour EDM scheduler si diffusers ne supporte pas (optional, Phase 3.2)

# --------------------------------------------------------------------------
# Development Tools (Optional)
# --------------------------------------------------------------------------
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting
mypy>=1.0.0  # Type checking
pre-commit>=3.0.0  # Git hooks

# --------------------------------------------------------------------------
# Notes d'Installation
# --------------------------------------------------------------------------
# 
# Installation complète (recommandée):
#   pip install -r requirements.txt
#
# Installation minimale (sans Jupyter):
#   pip install -r requirements.txt --no-deps
#   puis installer manuellement les dépendances core
#
# Pour CUDA (GPU):
#   1. Installer PyTorch avec CUDA:
#      pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#   2. Installer PyTorch Geometric avec CUDA:
#      pip install torch-geometric torch-scatter torch-sparse -f https://data.pyg.org/whl/torch-2.0.0+cu118.html
#   3. Installer le reste:
#      pip install -r requirements.txt
#
# Pour CPU uniquement:
#   pip install -r requirements.txt
#
# Résolution de problèmes:
#   - Si torch-geometric échoue: installer les wheels depuis https://data.pyg.org/whl/
#   - Si netcdf4 échoue: conda install netcdf4
#   - Si dask est lent: pip install dask[complete] --upgrade
#
# ============================================================================
